==> 1_Instacart_Consistency check and Merger of ORDERS PRODUCTS and PRIOR.ipynb <==
# NOTEBOOK 1 - Consistency Check of ORDERS, PRODUCTS, ORDERS_PRODUCTS_PRIOR and Merge of ORDERS with ORDERS_PRODUCTS_PRIOR

DATASETS:
- ORDERS - original dataset with 3.421.083 rows and 7 columns
- ORDERS_PRODUCTS_PRIOR - original dataset with 32.434.489 rows and 4 columns
- PRODUCTS - original dataset with 49.693 rows and 5 columns

CONTENTS:
1. Consitency check of the ORDERS dataset
a. Dropping columns
b. Renaming columns
c. Missing values
d. Data types
e. Duplicates
    
2. Consistency check of the PRODUCTS dataset
a. Missing values
b. Data types
c. Duplicates
    
3. Consistency check of ORDERS_PRODUCTS_PRIOR dataset
a. Missing values
b. Data types
c. Duplicates
    
4. Merging ORDERS with ORDERS_PRODUCTS_PRIOR



# Import libraries

import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy   

# 1. Consistency check of ORDERS dataset = df_ORDS

# Import original "orders.csv" dataset

df_ORDS = pd.read_csv(r'D:\CareerFoundry\Data Immersion\Task 4\Instacart Basket Analysis_2021-07-22\02 Data\Original Data\orders.csv')

# Check the numbers of rows and columns

df_ORDS.shape

df_ORDS.head()

df_ORDS.tail()

df_ORDS.info()

    1.a. Dropping columns

# Drop the "eval_set" column as it has no use for the analysis

df_ORDS.drop(columns = ['eval_set'], inplace = True)

    1.b. Renaming columns

# Rename column "order_dow" to give more intuitive meaning

df_ORDS.rename(columns = {'order_dow' : 'order_day_of_the_week'}, inplace = True)

# Check the two changes

df_ORDS.head()

    1.c. Missing values

# Missing values using "isnull" function

df_ORDS.isnull().sum()

# Look for missing values in the "days_since_prior_order"

df_ORDS['days_since_prior_order'].value_counts(dropna = False)

    1.d Data types

# Check data types

df_ORDS.dtypes

# Look for mixed-data types

for col in df_ORDS.columns.tolist():
  weird = (df_ORDS[[col]].applymap(type) != df_ORDS[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ORDS[weird]) > 0:
    print (col)

# Change "order_id" from integer into string

df_ORDS['order_id'] = df_ORDS['order_id'].astype('str')

df_ORDS['order_id'].dtype

# Change "user_id" data type

df_ORDS['user_id'] = df_ORDS['user_id'].astype('str')

df_ORDS['user_id'].dtype

# String types data are excluded from the summary stats

df_ORDS.describe()

    1.e. Duplicates

# Look for duplicates

df_ORDS_dups = df_ORDS[df_ORDS.duplicated()]

# No duplicates found

# Export the wrangled dataset

df_ORDS.to_csv(r'D:\CareerFoundry\Data Immersion\Task 4\Instacart Basket Analysis_2021-07-22\02 Data\Prepared Data\df_ORDS_wrangled_final.csv', index=False )

# 2. Consistency check of PRODUCT dataset = df_PRODS

# Import the original "products.csv" dataset

df_PRODS = pd.read_csv(r'D:\CareerFoundry\Data Immersion\Task 4\Instacart Basket Analysis_2021-07-22\02 Data\Original Data\products.csv')

df_PRODS.head()

df_PRODS.tail()

df_PRODS.info()

    2.a. Missing values

# Look for missing values

df_PRODS.isnull().sum()

# Check the "product_name" column

df_PRODS['product_name'].value_counts(dropna = False)

    2.b. Data types

# Change "product_id", "aisle_id", "department_id" data type into string

df_PRODS['product_id'] = df_PRODS['product_id'].astype('str')
df_PRODS['aisle_id'] = df_PRODS['aisle_id'].astype('str')
df_PRODS['department_id'] = df_PRODS['department_id'].astype('str')

# Check the change

df_PRODS.info()

# Look for mixed-data types

for col in df_PRODS.columns.tolist():
  weird = (df_PRODS[[col]].applymap(type) != df_PRODS[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_PRODS[weird]) > 0:
    print (col)
    
# The result shows "product_name" column because it contains 16 empty cells NaN

# Delete 16 rows with missing values in the "product_name" columns

df_PRODS_clean = df_PRODS[df_PRODS['product_name'].isnull() == False]

df_PRODS_clean.info()

df_PRODS_clean.describe()

2.c. Duplicates

# Look for duplicates

df_PRODS_dups = df_PRODS_clean[df_PRODS_clean.duplicated()]

# No duplicates found

# Export the wrangled dataset

df_PRODS_clean.to_csv(r'D:\CareerFoundry\Data Immersion\Task 4\Instacart Basket Analysis_2021-07-22\02 Data\Prepared Data\df_PRODS_wrangled_final.csv', index=False )

# 3. Consistency check of ORDERS_PRODUCTS_PRIOR
This data set contains data from the past and need to be merged with "orders" set to inclued "product_id" in the "orders" set

# Import the original dataset

df_ORDS_PRODS_prior = pd.read_csv(r'D:\CareerFoundry\Data Immersion\Task 4\Instacart Basket Analysis_2021-07-22\02 Data\Original Data\order_products__prior.csv')

df_ORDS_PRODS_prior.head()

df_ORDS_PRODS_prior.tail()

df_ORDS_PRODS_prior.info()

    3.a. Missing values

df_ORDS_PRODS_prior.isnull().sum()

    3.b. Data types

for col in df_ORDS_PRODS_prior.columns.tolist():
  weird = (df_ORDS_PRODS_prior[[col]].applymap(type) != df_ORDS_PRODS_prior[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ORDS_PRODS_prior[weird]) > 0:
    print (col)

# Change "product_id", "aisle_id", "department_id" data type

df_ORDS_PRODS_prior['product_id'] = df_ORDS_PRODS_prior['product_id'].astype('str')
df_ORDS_PRODS_prior['order_id'] = df_ORDS_PRODS_prior['order_id'].astype('str')


    3.c. Duplicates

df_ORDS_PRODS_prior_dups = df_ORDS_PRODS_prior[df_ORDS_PRODS_prior.duplicated()]

# 4. Merging ORDERS with ORDERS_PRODUCTS_PRIOR
The merge will be performed on "order_id" variable; data type has been unified in both datasets to string

df_ORDS.info()

df_ORDS_PRODS_prior.info()

# merging the two data sets on column "order_id"
# the order set has  3.421.083 rows
# the prior set has 32.434.489 rows

# The output below shows the frequency of a merge using the argument how = outer. 
# Merging like this will combine all the observations and show the real merge rate:

df_ORDS_large = df_ORDS.merge(df_ORDS_PRODS_prior, on = 'order_id', how = 'outer', indicator = True)

df_ORDS_large.info()

# the merged dataset contains 32.640.698 rows and 10 columns

df_ORDS_large.head()

df_ORDS_large['_merge'].value_counts()

# Drop the "_merge" column as it will prevent me from using the indicator during next merge

df_ORDS_large.drop(columns = ['_merge'], inplace = True)

df_ORDS_large.to_csv(r'D:\CareerFoundry\Data Immersion\Task 4\Instacart Basket Analysis_2021-07-22\02 Data\Prepared Data\df_ORDS_large_wrangled_final.csv', index=False)

df_ORDS_large.to_pickle(r'D:\CareerFoundry\Data Immersion\Task 4\Instacart Basket Analysis_2021-07-22\02 Data\Prepared Data\df_ORDS_large_wrangled_final.pkl')

# For continuation refer to notebook 2_Final_Instacart
