==> 2_Instacart_Merger ORDS_large with PRODS.ipynb <==
# NOTEBOOK 2 - Merger of ORDERS_PRODUCTS with PRODUCTS

DATASETS:
- ORDERS_PRODUCTS = df_ORDS_large - merged dataset with 32.640.698 rows and 9 columns
- PRODUCTS = df_PRODS_clean - wrangled dataset with 49.677 rows and 5 columns

CONTENTS:
1. Data types
2. Merge
3. Missing values

# Import libraries

import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

# For the merge purposes I am working on csv files

df_ORDS_large = pd.read_csv(r'D:\CareerFoundry\Data Immersion\Task 4\Instacart Basket Analysis_2021-07-22\02 Data\Prepared Data\df_ORDS_large_wrangled_final.csv')

df_PRODS_clean = pd.read_csv(r'D:\CareerFoundry\Data Immersion\Task 4\Instacart Basket Analysis_2021-07-22\02 Data\Prepared Data\df_PRODS_wrangled_final.csv')

    1. Data types

df_ORDS_large.info()

df_PRODS_clean.info()

# Change "product_id" data type into float

df_PRODS_clean['product_id'] = df_PRODS_clean['product_id'].astype('float64')

df_PRODS_clean.info()

    2. Merge

# Merge with df_ORDS_large

df_ORDS_PRODS_merged = df_ORDS_large.merge(df_PRODS_clean, on = 'product_id', how = 'outer', indicator = True)

df_ORDS_PRODS_merged.info()

df_ORDS_PRODS_merged['_merge'].value_counts()

df_ORDS_PRODS_merged.drop(columns = ['_merge'], inplace = True)

df_ORDS_PRODS_merged.head()

    3. Missing values

df_ORDS_PRODS_merged.isnull().sum()

# Check the missing values in "order_id" column

df_NaN_order_id = df_ORDS_PRODS_merged[df_ORDS_PRODS_merged['order_id'].isnull() == True]

df_NaN_order_id

# it looks like those 11 products have never been ordered by any user yet

# Check the missing values in "days_since_prior_order" using crosstab as most likely those relate to all the first orders

crosstab_DSPO = pd.crosstab(df_ORDS_PRODS_merged['days_since_prior_order'], df_ORDS_PRODS_merged['order_number'], dropna = False)

crosstab_DSPO.to_clipboard()

# I am not sure if I am correct but in a dataset with 32.642.461 rows, where each rows represent a product attached to each order_id and user_id, 
# 2.078.218 products can be labeled as sold within first order of each user because this is the number of missing values in "days_since_prior_order". 

df_ORDS_PRODS_merged.to_csv(r'D:\CareerFoundry\Data Immersion\Task 4\Instacart Basket Analysis_2021-07-22\02 Data\Prepared Data\df_ORDS_PRODS_merged.csv', index = False)

df_ORDS_PRODS_merged.to_pickle(r'D:\CareerFoundry\Data Immersion\Task 4\Instacart Basket Analysis_2021-07-22\02 Data\Prepared Data\df_ORDS_PRODS_merged.pkl')

# For continuation refer to notebook 3_Final_Instacart
