==> 3_Instacart_Merger with CUSTOMERS.ipynb <==
# NOTEBOOK 3 - Merger of ORDERS_PRODUCTS with CUSTOMERS

DATASETS:
- ORDERS_PRODUCTS = df_ORDS_PRODS - merged dataset with 32.640.698 rows and 9 columns
- CUSTOMERS = df_CUST - original dataset with 206.209 rows and 10 columns

CONTENTS:
1. Consistency check
2. Merge

import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

# Import the original dataset with CUSTOMERS

df_CUST = pd.read_csv(r'D:\CareerFoundry\Data Immersion\Task 4\Instacart Basket Analysis_2021-07-22\02 Data\Original Data\customers.csv')

# Import the prepared ORDS_PRODS dataset

df_ORDS_PRODS = pd.read_pickle(r'D:\CareerFoundry\Data Immersion\Task 4\Instacart Basket Analysis_2021-07-22\02 Data\Prepared Data\df_ORDS_PRODS_merged.pkl')

1. Consistency check

df_CUST.info()

# there are 206.209 rows with customers data which equals to 206.209 missing values from the column "days_since_prior_order" in the ORDERS set, which confirms that those rows are related to all first orders

# Rename columns

df_CUST.rename(columns = {'First Name' : 'first_name', 'Surnam' : 'last_name', 'Gender' : 'gender', 'STATE' : 'state', 'Age' : 'age' }, inplace = True)

df_CUST.head()

# Check for missing values

df_CUST.isnull().sum()

For now, I leave the missing values unaddressed. I task 4.10 the "first_name" and "last_name" column will be deleted due to privacy concerns

# Check for strange values

df_CUST.describe()

df_ORDS_PRODS.info()

df_CUST.info()

# "user_id" is the common variable in both set; need to make sure the variable is the same data type

df_CUST['user_id'] = df_CUST['user_id'].astype('float64')

df_CUST.info()

2. Merger

df_CPO = df_ORDS_PRODS.merge(df_CUST, on = 'user_id')

# For the exercise sake I left out the "how" and the "indicator" arguments. Only this way I could successfully merge those sets

# The merged dataset

df_CPO.info()

df_CPO.isnull().sum()

df_CPO.to_pickle(r'D:\CareerFoundry\Data Immersion\Task 4\Instacart Basket Analysis_2021-07-22\02 Data\Prepared Data\df_CPO_merged.pkl')

END of MERGERS

# FOR BUSINESS QUESTIONS PLEASE REFER TO NOTEBOOK "Task 10 Coding etiquette & Excel reporting"
